chore: refactor naming and structure
