+++ src/main/java/service/validation/ValidationFactory.java
@@ -1,6 +1,6 @@
-package service.strategy;
+package service.validation;
 
-import service.validator.*;
+import service.validation.validator.*;
 import servlet.Route;
 
 import java.util.Collections;
@@ -8,26 +8,24 @@ import java.util.List;
 import java.util.Map;
 
 public class ValidationFactory {
-    private final Map<String, Map<String, List<FieldValidator>>> routesToValidationMap;
+    private final Map<String, Map<String, List<Validator>>> routesToValidationMap;
 
     public ValidationFactory() {
         routesToValidationMap = Map.of(
-                Route.TEST_REGISTRATION_PAGE.getUri(),Map.of(
-                        "password",List.of(
+                Route.TEST_REGISTRATION_PAGE.getUri(), Map.of(
+                        "password", List.of(
                                 new EmptinessValidator(),
                                 new LengthValidator(16),
                                 new SpecialCharacterValidator(2)),
 
                         "email", List.of(
                                 new EmptinessValidator(),
-                                new EmailValidator())
+                                new EmailPatternValidator())
                 )
         );
     }
 
-    public Map<String, List<FieldValidator>> getValidationsByContextPath(String contextPath) {
+    public Map<String, List<Validator>> getValidationsByContextPath(String contextPath) {
         return routesToValidationMap.getOrDefault(contextPath, Collections.emptyMap());
     }
-
-
 }
