+++ src/main/java/service/validation/validator/EmailPatternValidator.java
@@ -1,13 +1,13 @@
-package service.validator;
+package service.validation.validator;
 
 import java.util.Collections;
 import java.util.List;
 
-public class EmailValidator implements FieldValidator {
+public class EmailPatternValidator implements Validator {
     private final String emailPattern;
     private final List<String> errorMessages;
 
-    public EmailValidator() {
+    public EmailPatternValidator() {
         this.emailPattern = "^[A-Za-z0-9+_.-]+@(.+)$";
         this.errorMessages = List.of(
                 "Email doesn't match the pattern",
